+++ __tests__/jest-test.js
@@ -1,0 +1,5 @@
+const jestConfig = require('../jest');
+
+it('exports a valid config', () => {
+  expect(jestConfig).toBeTruthy();
+});